% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/baseline_calibration_methods_robust.R
\name{beta_calibration}
\alias{beta_calibration}
\title{Apply Beta Calibration - ROBUST}
\usage{
beta_calibration(train_probs, train_labels, test_probs)
}
\description{
Apply Beta Calibration - ROBUST
}
