% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calibrate_predictions.R
\name{calibrate_predictions}
\alias{calibrate_predictions}
\title{Recalibrate predicted probabilities using SFI}
\usage{
calibrate_predictions(
  df,
  sfi_reference,
  alpha = 1,
  prob_col = "predicted_prob"
)
}
\arguments{
\item{df}{A data.frame with columns: predicted_prob and SFI}

\item{sfi_reference}{Mean SFI from training dataset}

\item{alpha}{Sensitivity multiplier (default = 1.0)}

\item{prob_col}{Name of the column with original model probabilities}
}
\value{
A data.frame with an added column 'calibrated_prob'
}
\description{
Recalibrate predicted probabilities using SFI
}
