% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dynamic_alpha_tuning.R
\name{dynamic_alpha_selection}
\alias{dynamic_alpha_selection}
\title{Dynamically Select Optimal Alpha Using Cross-Validation}
\usage{
dynamic_alpha_selection(
  df_train,
  sfi_reference,
  alpha_range = seq(0.5, 3, by = 0.1),
  n_folds = 5,
  metric = "brier"
)
}
\arguments{
\item{df_train}{Training data with SFI scores and predictions}

\item{sfi_reference}{Reference SFI mean}

\item{alpha_range}{Range of alpha values to test}

\item{n_folds}{Number of CV folds}

\item{metric}{Optimization metric ("brier", "ece", "auc")}
}
\description{
Dynamically Select Optimal Alpha Using Cross-Validation
}
