% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brier_decomposition.R
\name{expected_calibration_error}
\alias{expected_calibration_error}
\title{Compute Expected Calibration Error (ECE)}
\usage{
expected_calibration_error(labels, predictions, n_bins = 10)
}
\arguments{
\item{labels}{Binary outcome vector}

\item{predictions}{Predicted probabilities}

\item{n_bins}{Number of bins (default = 10)}
}
\description{
Compute Expected Calibration Error (ECE)
}
