% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sfi.R
\name{find_2d_peaks}
\alias{find_2d_peaks}
\title{Feature extraction core function}
\usage{
find_2d_peaks(
  mz,
  rt,
  intensity,
  ppm = 5,
  deltart = 5,
  snr = 3,
  mz_bins = 50000,
  rt_bins = 100
)
}
\arguments{
\item{mz}{Numeric vector of m/z values.}

\item{rt}{Numeric vector of retention times.}

\item{intensity}{Numeric vector of intensities corresponding to m/z and rt values.}

\item{ppm}{Numeric. Parts per million tolerance for m/z matching. Default is 5.}

\item{deltart}{Numeric. Tolerance for retention time matching. Default is 5.}

\item{snr}{Numeric. signal to ratio to find peaks.}

\item{mz_bins}{Numeric. m/z bins. Default 50000.}

\item{rt_bins}{Numeric. retention time bins. Default 100.}
}
\value{
A data frame containing the aligned and filtered Quality Control Matrix.
}
\description{
This function finds local max peaks on m/z-retention time 2D plane.
}
\examples{
data(sfi)
peak <- find_2d_peaks(mz=sfi$mz,rt=sfi$rt,intensity=sfi$intensity)
}
