% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/low_res.R
\name{find_peaks_low_res}
\alias{find_peaks_low_res}
\title{Find peaks in low-resolution data using the 2D peak finding algorithm}
\usage{
find_peaks_low_res(mz, rt, intensity, deltart = 5, snr = 3)
}
\arguments{
\item{mz}{A numeric vector of mass-to-charge ratios.}

\item{rt}{A numeric vector of retention times.}

\item{intensity}{A numeric vector of intensities.}

\item{deltart}{Numeric. Tolerance for retention time matching. Default is 5.}

\item{snr}{Numeric. Signal-to-noise ratio to find peaks. Default is 3.0.}
}
\value{
A data frame with columns `mz`, `rt`, and `intensity`,
        representing the detected peaks. The `mz` values are integer masses.
}
\description{
This function adapts the fast `find_2d_peaks` function for use with
low-resolution (unit mass) data. It does this by first aggregating the
signal at each integer mass and then calling the 2D peak finder.
}
\examples{
data(sfi)
peaks <- find_peaks_low_res(
    mz = sfi$mz, rt = sfi$rt,
    intensity = sfi$intensity
)
}
