% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_all.R
\name{generate_all}
\alias{generate_all}
\title{Generate all visualizations}
\usage{
generate_all(output_dir = NULL, date = NULL)
}
\arguments{
\item{output_dir}{The directory to which the files should be written. If
\code{NULL} (the default), the current working directory will be used.}

\item{date}{A date to be printed on the report. If \code{NULL} (the
default), the current date will be used}
}
\value{
new files written
}
\description{
Generates an HTML, PDF, and individual plot files.
}
