% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_race_distribution.R
\name{generate_race_distribution}
\alias{generate_race_distribution}
\title{Generate Race Distribution}
\usage{
generate_race_distribution()
}
\value{
A named numeric vector of race proportions that sum to 1.
}
\description{
This function generates a random race distribution (White, Black, Hispanic, Asian, Other)
based on uniform sampling within specified ranges and ensures that "Other" remains
between 0 and 0.03.
}
\examples{
generate_race_distribution()

}
