% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sfi.R
\name{get_qc_features}
\alias{get_qc_features}
\alias{get_qc_features.sfi_peaks}
\alias{get_qc_features.default}
\title{Generate Quality Control Feature List}
\usage{
get_qc_features(mz, rt, intensity, ...)

\method{get_qc_features}{sfi_peaks}(mz, rt = NULL, intensity = NULL, ...)

\method{get_qc_features}{default}(
  mz,
  rt,
  intensity,
  idelta = 60,
  windows = 600,
  qcseq = c(1, 1, 0, 1, 1, 0, 1, 1, 0),
  deltart = 5,
  ppm = 5,
  minn = 6,
  ...
)
}
\arguments{
\item{mz}{Numeric vector of m/z values or an object of class `sfi_peaks`.}

\item{rt}{Numeric vector of retention times.}

\item{intensity}{Numeric vector of intensities corresponding to m/z and rt values.}

\item{...}{Additional arguments passed to methods.}

\item{idelta}{Numeric. Initial delta retention time. Default is 60.}

\item{windows}{Numeric. Retention time window. Default is 600.}

\item{qcseq}{Integer vector indicating QC samples. Default is c(1, 1, 0, 1, 1, 0, 1, 1, 0).}

\item{deltart}{Numeric. Tolerance for retention time matching. Default is 5.}

\item{ppm}{Numeric. Parts per million tolerance for m/z matching. Default is 5.}

\item{minn}{Integer. Minimum number of QC samples required. Default is 6.}
}
\value{
A data frame containing filtered QC features with the following columns:
\itemize{
  \item mzqc: aligned m/z of the QC feature.
  \item rtqc: aligned retention time of the QC feature.
  \item intensity: intensity of the feature in the specific QC sample.
  \item sampleidx: index of the QC sample injection.
  \item idxq: unique identifier for the QC feature group (mz rt).
}
The row names of the data frame are set to the sample index (injection number), with suffixes to ensure uniqueness.
}
\description{
This function generates a list of features found in Quality Control (QC) samples by aligning QC and matrix samples
and filtering based on detection frequency criteria.
}
\section{Methods (by class)}{
\itemize{
\item \code{get_qc_features(sfi_peaks)}: Method for sfi_peaks object

\item \code{get_qc_features(default)}: Default method for vectors

}}
\examples{
data(sfi)
peak <- find_2d_peaks(mz=sfi$mz,rt=sfi$rt,intensity=sfi$intensity)
qc_features <- get_qc_features(peak$mz, peak$rt, peak$intensity,
                               idelta=92.25,windows=632.11,minn=6,deltart=10)
}
