% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sfi.R
\name{getidelta}
\alias{getidelta}
\title{Optimize Delta Retention Time}
\usage{
getidelta(
  mz,
  rt,
  qcmz,
  qcrt,
  idelta = 60,
  shift = 0,
  ppm = 5,
  deltart = 5,
  window = 600,
  n = 160,
  tol = 0.03,
  max_iter = 100
)
}
\arguments{
\item{mz}{Numeric vector of m/z values.}

\item{rt}{Numeric vector of retention times.}

\item{qcmz}{Numeric vector of QC m/z values.}

\item{qcrt}{Numeric vector of QC retention times.}

\item{idelta}{Initial delta retention time guess. Default is 60.}

\item{shift}{Numeric. Shift applied to idelta. Default is 0.}

\item{ppm}{Numeric. Parts per million tolerance for m/z matching. Default is 5.}

\item{deltart}{Numeric. Tolerance for retention time matching. Default is 5.}

\item{window}{Numeric. Retention time window. Default is 600.}

\item{n}{Integer. Number of iterations or samples. Default is 160.}

\item{tol}{Numeric. Tolerance for binary search convergence. Default is 0.03.}

\item{max_iter}{Integer. Maximum number of binary search iterations. Default is 100.}
}
\value{
Optimized delta retention time (idelta).
}
\description{
This function optimizes the delta retention time (idelta) using a binary search approach.
}
\examples{
data(sfi)
peak <- find_2d_peaks(mz=sfi$mz,rt=sfi$rt,intensity=sfi$intensity)
peaklist2 <- getidelta(peak$mz, peak$rt,qcmz=195.0876,qcrt=74,window=632,idelta=90)
}
