% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sfi.R
\name{getqc}
\alias{getqc}
\title{Quality Control for Mass Spectrometry Data}
\usage{
getqc(
  mz,
  rt,
  intensity,
  idelta = 60,
  window = 600,
  qcseq = c(1, 1, 0, 1, 1, 0, 1, 1, 0),
  deltart = 5,
  ppm = 5,
  minn = 1,
  n = 160,
  tol = 0.03,
  max_iter = 100,
  wlower = 620,
  wupper = 650
)
}
\arguments{
\item{mz}{Numeric vector of m/z values.}

\item{rt}{Numeric vector of retention times.}

\item{intensity}{Numeric vector of intensities corresponding to m/z and rt values.}

\item{idelta}{Numeric. Initial delta retention time. Default is 60.}

\item{window}{Numeric. Retention time window. Default is 600.}

\item{qcseq}{Integer vector indicating QC samples. Default is c(1, 1, 0, 1, 1, 0, 1, 1, 0).}

\item{deltart}{Numeric. Tolerance for retention time matching. Default is 5.}

\item{ppm}{Numeric. Parts per million tolerance for m/z matching. Default is 5.}

\item{minn}{Integer. Minimum number of QC samples required. Default is 1.}

\item{n}{Integer. Number of samples for delta optimization. Default is 160.}

\item{tol}{Numeric. Tolerance for binary search in delta optimization. Default is 0.03.}

\item{max_iter}{Integer. Maximum iterations for binary search. Default is 100.}

\item{wlower}{Numeric. Lower bound for window determination. Default is 620.}

\item{wupper}{Numeric. Upper bound for window determination. Default is 650.}
}
\value{
A numeric vector containing the optimal window and delta retention time.
}
\description{
This function performs quality control (QC) on mass spectrometry data by aligning QC and sample features.
}
