% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sfi.R
\name{getqcdf}
\alias{getqcdf}
\title{Generate Quality Control Data Frame}
\usage{
getqcdf(
  mz,
  rt,
  intensity,
  idelta = 60,
  windows = 600,
  qcseq = c(1, 1, 0, 1, 1, 0, 1, 1, 0),
  deltart = 5,
  ppm = 5,
  minn = 6
)
}
\arguments{
\item{mz}{Numeric vector of m/z values.}

\item{rt}{Numeric vector of retention times.}

\item{intensity}{Numeric vector of intensities corresponding to m/z and rt values.}

\item{idelta}{Numeric. Initial delta retention time. Default is 60.}

\item{windows}{Numeric. Retention time window. Default is 600.}

\item{qcseq}{Integer vector indicating QC samples. Default is c(1, 1, 0, 1, 1, 0, 1, 1, 0).}

\item{deltart}{Numeric. Tolerance for retention time matching. Default is 5.}

\item{ppm}{Numeric. Parts per million tolerance for m/z matching. Default is 5.}

\item{minn}{Integer. Minimum number of QC samples required. Default is 6.}
}
\value{
A data frame containing QC information after alignment and filtering.
}
\description{
This function generates a QC data frame by aligning QC and matrix samples and filtering based on criteria.
}
\examples{
data(sfi)
peak <- find_2d_peaks(mz=sfi$mz,rt=sfi$rt,intensity=sfi$intensity)
peaklist <- getqcdf(peak$mz, peak$rt, peak$intensity,idelta=92.25,windows=632.11,minn=6,deltart=10)
}
