% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sfi.R
\name{getwindow}
\alias{getwindow}
\alias{getwindow.sfi_peaks}
\alias{getwindow.default}
\title{Determine Optimal Retention Time Window}
\usage{
getwindow(mz, rt, ...)

\method{getwindow}{sfi_peaks}(mz, rt = NULL, ...)

\method{getwindow}{default}(
  mz,
  rt,
  lower = 620,
  upper = 650,
  ppm = 5,
  minn = 1,
  qcseq = c(1, 1, 0, 1, 1, 0, 1, 1, 0),
  ...
)
}
\arguments{
\item{mz}{Numeric vector of m/z values or an object of class `sfi_peaks`.}

\item{rt}{Numeric vector of retention times.}

\item{...}{Additional arguments passed to methods.}

\item{lower}{Numeric. Lower bound for the retention time window. Default is 620.}

\item{upper}{Numeric. Upper bound for the retention time window. Default is 650.}

\item{ppm}{Numeric. Parts per million tolerance for m/z matching. Default is 5.}

\item{minn}{Integer. Minimum number of features in a QC cluster. Default is 1.}

\item{qcseq}{Integer vector. QC sequence indicating which samples are QC. Default is c(1, 1, 0, 1, 1, 0, 1, 1, 0).}
}
\value{
Numeric value representing the optimal retention time window.
}
\description{
This function calculates the optimal retention time window based on QC sequences and m/z/rt data.
}
\section{Methods (by class)}{
\itemize{
\item \code{getwindow(sfi_peaks)}: Method for sfi_peaks object

\item \code{getwindow(default)}: Default method for vectors

}}
\examples{
data(sfi)
peak <- find_2d_peaks(mz=sfi$mz,rt=sfi$rt,intensity=sfi$intensity)
window_opt <- getwindow(peak$mz, peak$rt)
}
