% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/baseline_calibration_methods_robust.R
\name{histogram_binning}
\alias{histogram_binning}
\title{Apply Histogram Binning Calibration - ROBUST}
\usage{
histogram_binning(train_probs, train_labels, test_probs, n_bins = 10)
}
\description{
Apply Histogram Binning Calibration - ROBUST
}
