% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/baseline_calibration_methods_robust.R
\name{isotonic_calibration}
\alias{isotonic_calibration}
\title{Apply Isotonic Regression Calibration - ROBUST}
\usage{
isotonic_calibration(train_probs, train_labels, test_probs)
}
\description{
Apply Isotonic Regression Calibration - ROBUST
}
