% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/identify_optimal_alpha.R
\name{plot.sfi_optimal_alpha}
\alias{plot.sfi_optimal_alpha}
\title{Plot Method for sfi_optimal_alpha Objects}
\usage{
\method{plot}{sfi_optimal_alpha}(x, ...)
}
\arguments{
\item{x}{An object of class sfi_optimal_alpha}

\item{...}{Additional arguments passed to print.ggplot}
}
\description{
Plot Method for sfi_optimal_alpha Objects
}
