% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/identify_optimal_alpha.R
\name{print.sfi_optimal_alpha}
\alias{print.sfi_optimal_alpha}
\title{Print Method for sfi_optimal_alpha Objects}
\usage{
\method{print}{sfi_optimal_alpha}(x, ...)
}
\arguments{
\item{x}{An object of class sfi_optimal_alpha}

\item{...}{Additional arguments (not used)}
}
\description{
Print Method for sfi_optimal_alpha Objects
}
