% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sfi.R
\docType{data}
\name{sfi}
\alias{sfi}
\title{Demo sfi data}
\format{
A data.frame object with mass to charge ratio, intensity and retention time from sfi mode.
}
\usage{
data(sfi)
}
\description{
Demo sfi data
}
\keyword{datasets}
