% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sfi_plot.R
\name{sfi_plot}
\alias{sfi_plot}
\title{SFI plot}
\usage{
sfi_plot(author = "Joe", figure = 1, save = FALSE)
}
\arguments{
\item{author}{Character. The last name of the publishing author.}

\item{figure}{Integer. The figure number in the corresponding paper.}

\item{save}{Boolean. Whether to save as a file or simple return the plot (the default).}
}
\value{
a plot object.
}
\description{
Generate a plot for the SFI Law as Data Project, specfic to each Author and figure type.
}
