% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/baseline_calibration_methods_robust.R
\name{temperature_scaling}
\alias{temperature_scaling}
\title{Apply Temperature Scaling - ROBUST}
\usage{
temperature_scaling(train_probs, train_labels, test_probs)
}
\description{
Apply Temperature Scaling - ROBUST
}
