% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_sfi.R
\name{theme_sfi}
\alias{theme_sfi}
\title{SFI visualization theme}
\usage{
theme_sfi(base_size = 12, base_family = "CMU Bright",
  nomargin = FALSE, fc = "black", gM = TRUE, gm = FALSE,
  gc = "grey", gl = "dashed", boxes = FALSE, bc = "white",
  pc = "transparent", lp = "right", ld = NULL, lt = 0.8,
  lj = NULL, lba = FALSE, la = NULL, lfb = "plain", lkh = FALSE,
  lkw = FALSE, lkt = NULL, legend_height = NULL,
  legend_width = NULL, axis = 1)
}
\arguments{
\item{base_size}{The font size}

\item{base_family}{The font}

\item{nomargin}{The margin}

\item{fc}{The font color}

\item{gM}{Major grid}

\item{gm}{Minor grid}

\item{gc}{Color of grid}

\item{gl}{Line type}

\item{boxes}{Box around plot}

\item{bc}{Background color}

\item{pc}{Panel background color}

\item{lp}{Legend position}

\item{ld}{Legen direction}

\item{lj}{legend justification}

\item{lba}{legend background}

\item{lfb}{legend font bold}

\item{lkh}{legend height}

\item{lkw}{legend width}

\item{legend_height}{numeric height}

\item{legend_width}{numeric width}

\item{axis}{Axis angle}
}
\value{
A ggplot2 theme
}
\description{
The Santa Fe Institute Law as Data visualization theme
}
