% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sfi.R
\name{getmzml}
\alias{getmzml}
\title{Read mzML File and Extract m/z, Retention Time, and Intensity}
\usage{
getmzml(path)
}
\arguments{
\item{path}{path of SFI mzML file.}
}
\value{
A matrix containing m/z, retention time and intensity.
}
\description{
Read mzML File and Extract m/z, Retention Time, and Intensity
}
