% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sfi.R
\name{getsff}
\alias{getsff}
\title{Cluster and Pair m/z and Retention Time Features}
\usage{
getsff(mz, rt, ppm = 5, minn = 2, refmz = NULL)
}
\arguments{
\item{mz}{Numeric vector of m/z values.}

\item{rt}{Numeric vector of retention times corresponding to m/z values.}

\item{ppm}{Numeric. Parts per million tolerance for m/z matching. Default is 5.}

\item{minn}{Integer. Minimum number of features in a cluster to be retained. Default is 2.}

\item{refmz}{Optional numeric vector of reference m/z values for alignment. Default is NULL.}
}
\value{
A data frame containing paired m/z and retention time values with their differences.
}
\description{
This function clusters m/z values based on Manhattan distance and pairs features within clusters.
}
