% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sfi.R
\name{getsfm}
\alias{getsfm}
\title{Generate Quality Control Matrix (SFM)}
\usage{
getsfm(
  mz,
  rt,
  intensity,
  idelta = 60,
  windows = 600,
  qcseq = c(1, 1, 0, 1, 1, 0, 1, 1, 0),
  deltart = 5,
  ppm = 5,
  minn = 1,
  n = 160
)
}
\arguments{
\item{mz}{Numeric vector of m/z values.}

\item{rt}{Numeric vector of retention times.}

\item{intensity}{Numeric vector of intensities corresponding to m/z and rt values.}

\item{idelta}{Numeric. Initial delta retention time. Default is 60.}

\item{windows}{Numeric. Retention time window. Default is 600.}

\item{qcseq}{Integer vector indicating QC samples. Default is c(1, 1, 0, 1, 1, 0, 1, 1, 0).}

\item{deltart}{Numeric. Tolerance for retention time matching. Default is 5.}

\item{ppm}{Numeric. Parts per million tolerance for m/z matching. Default is 5.}

\item{minn}{Integer. Minimum number of QC samples required. Default is 1.}

\item{n}{Integer. Number of samples for delta optimization. Default is 160.}
}
\value{
A data frame containing the aligned and filtered Quality Control Matrix.
}
\description{
This function generates a Quality Control Matrix by aligning and filtering sample and matrix peaks.
}
