% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sfi_plot.R
\name{sfi_plot}
\alias{sfi_plot}
\title{SFI plot}
\usage{
sfi_plot(author = "laquer", figure = "1")
}
\arguments{
\item{author}{Character. The last name of the publishing author.}

\item{figure}{Character. The figure number in the corresponding paper.}
}
\value{
a plot object.
}
\description{
Generate a plot for the SFI Law as Data Project, specfic to each Author and figure type.
}
