% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sfi_theme.R
\name{theme_sfi}
\alias{theme_sfi}
\title{SFI visualization theme}
\usage{
theme_sfi(base_size = 12, base_family = "CMU Bright Roman",
  nomargin = FALSE, fc = "black", gM = TRUE, gm = FALSE,
  gc = "grey", gl = "dashed", boxes = FALSE, bc = "white",
  pc = "transparent", lp = "right", axis = 1)
}
\arguments{
\item{base_size}{The font size}

\item{base_family}{The font}

\item{nomargin}{The margin}

\item{fc}{The font color}

\item{gM}{Major grid}

\item{gm}{Minor grid}

\item{gc}{Color of grid}

\item{gl}{Line type}

\item{boxes}{Box around plot}

\item{bc}{Background color}

\item{pc}{Panel background color}

\item{lp}{Legend position}

\item{axis}{Axis angle}
}
\value{
A ggplot2 theme
}
\description{
The Santa Fe Institute Law as Data visualization theme
}
